tau fixed but still diverges - perhaps need check on gap computation

+++ postProcessing.cpp
@@ -11,7 +11,7 @@ void export_result(const char *method, T *x, int numNodes){
 	// Write results to this file
 	FILE *f = fopen(FileName, "wb");
 	for (int i = 0; i < numNodes; i++){
-		fprintf(f, " %i %i\n", i, (int)x[i]);
+		fprintf(f, " %i %f\n", i, x[i]);
 	}
 	fclose(f);
 }
